+++ bib/internal/container/container.go
@@ -123,26 +123,6 @@ func (c *Container) ExecArgv() []string {
 	return []string{"podman", "exec", "-i", c.id}
 }
 
-// InitDNF initializes dnf in the container. This is necessary when
-// the caller wants to read the image's dnf repositories, but they are
-// not static, but rather configured by dnf dynamically. The primaru
-// use-case for this is RHEL and subscription-manager.
-//
-// The implementation is simple: We just run plain `dnf` in the
-// container so that the subscription-manager gets initialized. For
-// compatibility with both dnf and dnf5 we cannot just run "dnf" as
-// dnf5 will error and do nothing in this case. So we use "dnf check
-// --duplicates" as this is fast on both dnf4/dnf5 (just doing "dnf5
-// check" without arguments takes around 25s so that is not a great
-// option).
-func (c *Container) InitDNF() error {
-	if output, err := exec.Command("podman", "exec", c.id, "dnf", "check", "--duplicates").CombinedOutput(); err != nil {
-		return fmt.Errorf("initializing dnf in %s container failed: %w\noutput:\n%s", c.id, err, string(output))
-	}
-
-	return nil
-}
-
 // DefaultRootfsType returns the default rootfs type (e.g. "ext4") as
 // specified by the bootc container install configuration. An empty
 // string is valid and means the container sets no default.
